Only show Amount input if button supports amount parameter
PayPal now prevents the populating of the Amount field on its end via
URL parameter. This only applies to _s-xclick-type buttons, and was done
so as an added security measure. This restriction does not apply to
_donations-type buttons, however _donations-type buttons do not support
recurring payments. The former option is the better of the two. For more
information, see:
<https://stackoverflow.com/questions/18026250/paypal-trying-to-pass-an-amount-with-a-donation-button>